+++ packages/frontend/src/components/ProjectConnection/index.tsx
@@ -241,16 +241,10 @@ export const UpdateProjectConnection: FC<{ projectUuid: string }> = ({
 
 export const CreateProjectConnection: FC = () => {
     const history = useHistory();
-    const { user, health } = useApp();
+    const { user, health, activeJobIsRunning, activeJob } = useApp();
     const onError = useOnProjectError();
     const createMutation = useCreateMutation();
-    const {
-        isLoading: isSaving,
-        mutateAsync,
-        isIdle,
-        isSuccess,
-        data,
-    } = createMutation;
+    const { isLoading: isSaving, mutateAsync } = createMutation;
     const methods = useForm<ProjectConnectionForm>({
         shouldUnregister: true,
         defaultValues: {
@@ -275,6 +269,14 @@ export const CreateProjectConnection: FC = () => {
         });
     };
 
+    useEffect(() => {
+        if (activeJob?.jobResults?.projectUuid) {
+            history.push({
+                pathname: `/createProjectSettings/${activeJob?.jobResults?.projectUuid}`,
+            });
+        }
+    }, [activeJob, history]);
+
     return (
         <Form
             name="create_project"
@@ -285,36 +287,17 @@ export const CreateProjectConnection: FC = () => {
             <ProjectFormProvider>
                 <ProjectForm
                     showGeneralSettings={!health.data?.needsProject}
-                    disabled={isSaving || isSuccess}
+                    disabled={isSaving || !!activeJobIsRunning}
                     defaultType={health.data?.defaultProject?.type}
                 />
             </ProjectFormProvider>
-            {!isIdle && (
-                <ProjectStatusCallout
-                    style={{ marginBottom: '20px' }}
-                    mutation={createMutation}
-                />
-            )}
-            {isSuccess ? (
-                <Button
-                    intent={Intent.PRIMARY}
-                    text="Next"
-                    onClick={async () => {
-                        history.push({
-                            pathname: `/createProjectSettings/${data?.projectUuid}`,
-                        });
-                    }}
-                    style={{ float: 'right' }}
-                />
-            ) : (
-                <Button
-                    type="submit"
-                    intent={Intent.PRIMARY}
-                    text="Test & save connection"
-                    loading={isSaving}
-                    style={{ float: 'right' }}
-                />
-            )}
+            <Button
+                type="submit"
+                intent={Intent.PRIMARY}
+                text="Test and compile project"
+                loading={isSaving || activeJobIsRunning}
+                style={{ float: 'right' }}
+            />
         </Form>
     );
 };