+++ packages/frontend/src/components/common/ActionCardList/index.tsx
@@ -4,9 +4,10 @@ import Fuse from 'fuse.js';
 import React, { useMemo, useState } from 'react';
 import { UseMutationResult } from 'react-query';
 import AddTilesToDashboardModal from '../../SavedDashboards/AddTilesToDashboardModal';
-import { DeleteDashboardModal } from '../../SavedDashboards/DeleteDashboardModal';
+import EmptySavedChartsState from '../../SavedQueries/EmptySavedChartsState';
 import ActionCard from '../ActionCard';
 import { ActionModalProps, ActionTypeModal } from '../modal/ActionModal';
+import DeleteActionModal from '../modal/DeleteActionModal';
 import UpdateActionModal from '../modal/UpdateActionModal';
 import {
     ActionCardListWrapper,
@@ -126,7 +127,7 @@ const ActionCardList = <
             )}
             {actionState.actionType === ActionTypeModal.DELETE &&
                 actionState.data && (
-                    <DeleteDashboardModal
+                    <DeleteActionModal
                         isOpen={
                             actionState.actionType === ActionTypeModal.DELETE
                         }
@@ -137,6 +138,7 @@ const ActionCardList = <
                         }}
                         uuid={actionState.data.uuid}
                         name={actionState.data.name}
+                        isChart={!!isChart}
                     />
                 )}
 
@@ -155,7 +157,14 @@ const ActionCardList = <
 
             {dataList.length <= 0 && (
                 <NoIdealStateWrapper>
-                    <NonIdealState title="No results available" icon="search" />
+                    {isChart ? (
+                        <EmptySavedChartsState />
+                    ) : (
+                        <NonIdealState
+                            title="No results available"
+                            icon="search"
+                        />
+                    )}
                 </NoIdealStateWrapper>
             )}
         </ActionCardListWrapper>