+++ packages/frontend/src/hooks/dashboard/useDashboards.tsx
@@ -3,15 +3,22 @@ import { useQuery } from 'react-query';
 import { lightdashApi } from '../../api';
 import useQueryError from '../useQueryError';
 
-export const getDashboards = async (projectUuid: string, chartId?: string) => {
-    const queryChartId: string = chartId ? `?chartUuid=${chartId}` : '';
-
-    return lightdashApi<DashboardBasicDetails[]>({
-        url: `/projects/${projectUuid}/dashboards${queryChartId}`,
+const getDashboards = async (projectUuid: string) =>
+    lightdashApi<DashboardBasicDetails[]>({
+        url: `/projects/${projectUuid}/dashboards`,
         method: 'GET',
         body: undefined,
     });
-};
+
+const getDashboardsContainingChart = async (
+    projectUuid: string,
+    chartId: string,
+) =>
+    lightdashApi<DashboardBasicDetails[]>({
+        url: `/projects/${projectUuid}/dashboards?chartUuid=${chartId}`,
+        method: 'GET',
+        body: undefined,
+    });
 
 export const useDashboards = (projectUuid: string) => {
     const setErrorResponse = useQueryError();
@@ -19,6 +26,18 @@ export const useDashboards = (projectUuid: string) => {
         queryKey: ['dashboards', projectUuid],
         queryFn: () => getDashboards(projectUuid || ''),
         enabled: projectUuid !== undefined,
+        onError: (result) => setErrorResponse(result),
+    });
+};
+
+export const useDashboardsContainingChart = (
+    projectUuid: string,
+    chartId: string,
+) => {
+    const setErrorResponse = useQueryError();
+    return useQuery<DashboardBasicDetails[], ApiError>({
+        queryKey: ['dashboards-containing-chart', projectUuid, chartId],
+        queryFn: () => getDashboardsContainingChart(projectUuid, chartId),
         onError: (result) => setErrorResponse(result),
     });
 };