+++ release.config.js
@@ -32,6 +32,10 @@ module.exports = {
           'e2e': {
             'npmPublish': false,
             'pkgRoot': 'packages/e2e'
+          },
+          'cli': {
+            'npmPublish': false,
+            'pkgRoot': 'packages/cli'
           }
         }
       }
@@ -83,7 +87,9 @@ module.exports = {
           'packages/common/package.json',
           'packages/backend/package.json',
           'packages/frontend/package.json',
-          'packages/e2e/package.json'
+          'packages/e2e/package.json',
+          'packages/cli/package.json',
+
         ],
         'message': 'chore(release): ${nextRelease.version} [skip ci]\n\n${nextRelease.notes}',
       }